Fixed some bugs around the database aspect of the generic events

+++ src/SFA.DAS.Events.Database/StoredProcedures/GetGenericEventsSinceEvent.sql
@@ -1,6 +1,6 @@
 ﻿CREATE PROCEDURE [dbo].[GetGenericEventsSinceEvent]
 	@eventTypes [dbo].[eventType] Readonly,	
-	@fromEventId int,
+	@fromEventId bigint,
 	@offset int,
 	@pageSize int
 AS
